+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [Unreleased](https://github.com/rajatthareja/ReportBuilder/compare/v1.3...master)
+## [Unreleased](https://github.com/rajatthareja/ReportBuilder/compare/v1.4...master)
+
+## [1.4](https://github.com/rajatthareja/ReportBuilder/compare/v1.3...v1.4) - 2017-10-29
 ### Fixed
 - Config options Fixed
 - Document Fixed